Move logging functions from util.py to a new lib/log.py
This (finally) fully solves the circular import problem:

$ for i in lib.util lib.thread lib.log lib.zookeeper external.stacktracer; do python3 -c "import $i"; done
$
