+++ lib/util.py
@@ -18,36 +18,18 @@
 Various utilities for SCION functionality.
 """
 
-import logging
 import os
 import signal
 import time
 import traceback
+
 from lib.defines import TOPOLOGY_PATH
 from external.stacktracer import trace_start
 
 CERT_DIR = 'certificates'
 SIG_KEYS_DIR = 'signature_keys'
 ENC_KEYS_DIR = 'encryption_keys'
-TRACE_DIR = "../traces"
-
-
-class _StreamErrorHandler(logging.StreamHandler):
-    """
-    A logging StreamHandler that will exit the application if there's a logging
-    exception.
-
-    We don't try to use the normal logging system at this point because we
-    don't know if that's working at all. If it is (e.g. when the exception is a
-    formatting error), when we re-raise the exception, it'll get handled by the
-    normal process.
-    """
-    def handleError(self, record):
-        self.stream.write("Exception in logging module:\n")
-        for line in traceback.format_exc().split("\n"):
-            self.stream.write(line+"\n")
-        self.flush()
-        raise
+TRACE_DIR = '../traces'
 
 
 def _get_isd_prefix(isd_dir):
@@ -175,25 +157,10 @@ def update_dict(dictionary, key, values, elem_num=0):
     dictionary[key] = dictionary[key][-elem_num:]
 
 
-def init_logging(level=logging.DEBUG):
-    """
-    Configure logging for components (servers, routers, gateways).
-    """
-    logging.basicConfig(level=level,
-                        handlers=[_StreamErrorHandler()],
-                        format='%(asctime)s [%(levelname)s]\t%(message)s')
-
-def log_exception(msg, *args, level=logging.CRITICAL, **kwargs):
-    """
-    Properly format an exception before logging
-    """
-    logging.log(level, msg, *args, **kwargs)
-    for line in traceback.format_exc().split("\n"):
-        logging.log(level, line)
-
 def trace():
     path = os.path.join(TRACE_DIR,
-                        "%s.trace.html" % os.environ['SUPERVISOR_PROCESS_NAME'])
+                        "%s.trace.html" %
+                        os.environ['SUPERVISOR_PROCESS_NAME'])
     trace_start(path)
 
 def timed(limit):
