+++ sphinx-doc/lib/log.rst
@@ -1,0 +1,3 @@
+
+.. automodule:: lib.log
+   :members:
